frontend/streamlit_app/requirements.txt
streamlit>=1.37,<2.0
requests>=2.32,<3.0
pydeck>=0.9
pandas>=2.2
pillow>=10.0
